+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Bug fixes:
  * icpp: fix the emulation of the x64 sse2 instructions with relocation;
  * icpp: fix the crash when getting the absolute source path;
+ * icpp: fix missing the last value of script's argv;
  * icpp: do not interpret memory-operation x64 instructions if there's no relocation binding to;
 
 Improvements: